refactor(workspace): extract model types
