+++ crates/hir/Cargo.toml
@@ -19,3 +19,4 @@ tracing.workspace = true
 triomphe.workspace = true
 utils.workspace = true
 vfs.workspace = true
+workspace-model.workspace = true
