+++ crates/workspace-model/src/source_db.rs
@@ -1,0 +1,67 @@
+use vfs::VfsPath;
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
+pub enum SourceFileKind {
+    #[default]
+    SystemVerilog,
+    IncludeHeader,
+    LibraryMap,
+    ProjectManifest,
+}
+
+impl SourceFileKind {
+    pub fn from_path(path: &VfsPath) -> Self {
+        match path.name_and_extension() {
+            Some((name, Some(ext))) if name == "vide" && ext.eq_ignore_ascii_case("toml") => {
+                Self::ProjectManifest
+            }
+            Some((_, Some(ext))) if ext.eq_ignore_ascii_case("map") => Self::LibraryMap,
+            Some((_, Some(ext)))
+                if ["vh", "svh", "svi"].iter().any(|header| ext.eq_ignore_ascii_case(header)) =>
+            {
+                Self::IncludeHeader
+            }
+            _ => Self::SystemVerilog,
+        }
+    }
+
+    pub fn is_semantic_compilation_unit(self) -> bool {
+        matches!(self, Self::SystemVerilog | Self::LibraryMap)
+    }
+
+    pub fn is_slang_parse_unit(self) -> bool {
+        matches!(self, Self::SystemVerilog | Self::LibraryMap)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use vfs::VfsPath;
+
+    use super::*;
+
+    #[test]
+    fn include_headers_are_not_standalone_parse_diagnostic_units() {
+        let kind =
+            SourceFileKind::from_path(&VfsPath::new_virtual_path("/include/defs.svh".into()));
+
+        assert_eq!(kind, SourceFileKind::IncludeHeader);
+        assert!(!kind.is_slang_parse_unit());
+    }
+
+    #[test]
+    fn systemverilog_sources_remain_parse_diagnostic_units() {
+        let kind = SourceFileKind::from_path(&VfsPath::new_virtual_path("/rtl/top.sv".into()));
+
+        assert_eq!(kind, SourceFileKind::SystemVerilog);
+        assert!(kind.is_slang_parse_unit());
+    }
+
+    #[test]
+    fn project_manifests_are_not_slang_parse_diagnostic_units() {
+        let kind = SourceFileKind::from_path(&VfsPath::new_virtual_path("/root/vide.toml".into()));
+
+        assert_eq!(kind, SourceFileKind::ProjectManifest);
+        assert!(!kind.is_slang_parse_unit());
+    }
+}
